requirements.txt
<<<<<<< HEAD
fastapi
uvicorn
openai
python-dotenv
=======
fastapi
uvicorn
openai
python-dotenv
>>>>>>> feature/fix-ai-agent
